+++ flitchannel.cpp
@@ -83,7 +83,7 @@ void FlitChannel::SendFlit( Flit* flit ) {
   else 
     ++_idle;
 
-  while ( (_queue.size() > _delay) && (_queue.front() == 0) )
+  while ( (_queue.size() > (unsigned int)_delay) && (_queue.front() == 0) )
     _queue.pop( );
 
   _queue.push(flit);
