+++ routers/chaos_router.cpp
@@ -252,7 +252,7 @@ bool ChaosRouter::_InputReady( int input ) const
 
 bool ChaosRouter::_OutputFull( int out ) const
 {
-  return ( _output_frame[out].size( ) >= _buffer_size );
+  return ( _output_frame[out].size( ) >= (unsigned int)_buffer_size );
 }
 
 bool ChaosRouter::_OutputAvail( int out ) const
@@ -264,7 +264,7 @@ bool ChaosRouter::_OutputAvail( int out ) const
 
 bool ChaosRouter::_MultiQueueFull( int mq ) const
 {
-  return ( _multi_queue[mq].size( ) >= _buffer_size );
+  return ( _multi_queue[mq].size( ) >= (unsigned int)_buffer_size );
 }
 
 int ChaosRouter::_InputForOutput( int output ) const
